api/requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
aiofiles
email-validator

pydantic[email]

requests
python-multipart